rename pointer_grab->cursor-grab
and grab owns the cursor
