+++ include/zen/input/pointer.h
@@ -2,14 +2,9 @@
 #define ZEN_POINTER_H
 
 #include <wayland-server.h>
-#include <wlr/interfaces/wlr_pointer.h>
-
-#include "zen/input/pointer_grab.h"
+#include <wlr/types/wlr_input_device.h>
 
 struct zn_pointer {
-  struct zn_pointer_grab* grab;
-  struct zn_pointer_grab grab_default;
-
   struct wl_listener button_listener;
   struct wl_listener motion_listener;
   struct wl_listener axis_listener;
