+++ zen/cursor.c
@@ -8,8 +8,65 @@
 #include <wlr/xcursor.h>
 
 #include "zen-common.h"
+#include "zen/input/cursor-grab.h"
 #include "zen/scene/screen-layout.h"
+#include "zen/scene/view.h"
 #include "zen/server.h"
+
+static void
+default_grab_motion(
+    struct zn_cursor_grab* grab, struct wlr_event_pointer_motion* event)
+{
+  UNUSED(grab);
+  struct zn_server* server = zn_server_get_singleton();
+  struct zn_cursor* cursor = server->input_manager->seat->cursor;
+  struct wlr_seat* seat = server->input_manager->seat->wlr_seat;
+  struct wlr_surface* surface;
+  struct zn_view* view;
+  double view_x, view_y;
+
+  view = zn_screen_get_view_at(
+      cursor->screen, cursor->x, cursor->y, &view_x, &view_y);
+
+  if (view != NULL) {
+    surface = view->impl->get_wlr_surface(view);
+  } else {
+    surface = NULL;
+  }
+
+  if (surface) {
+    wlr_seat_pointer_enter(seat, surface, view_x, view_y);
+    wlr_seat_pointer_send_motion(seat, event->time_msec, view_x, view_y);
+  } else {
+    zn_cursor_reset_surface(cursor);
+    wlr_seat_pointer_clear_focus(seat);
+  }
+}
+
+static void
+default_grab_button(
+    struct zn_cursor_grab* grab, struct wlr_event_pointer_button* event)
+{
+  UNUSED(grab);
+  struct zn_server* server = zn_server_get_singleton();
+  struct wlr_seat* seat = server->input_manager->seat->wlr_seat;
+  struct zn_cursor* cursor = server->input_manager->seat->cursor;
+  struct zn_view* view;
+
+  wlr_seat_pointer_send_button(
+      seat, event->time_msec, event->button, event->state);
+
+  if (event->state == WLR_BUTTON_PRESSED) {
+    view =
+        zn_screen_get_view_at(cursor->screen, cursor->x, cursor->y, NULL, NULL);
+    zn_scene_set_focused_view(server->scene, view);
+  }
+}
+
+static const struct zn_cursor_grab_interface default_grab_interface = {
+    .motion = default_grab_motion,
+    .button = default_grab_button,
+};
 
 static void
 zn_cursor_update_size(struct zn_cursor* self)
@@ -263,6 +320,9 @@ zn_cursor_create(void)
   self->surface_destroy_listener.notify = zn_cursor_handle_surface_destroy;
   wl_list_init(&self->surface_destroy_listener.link);
 
+  self->grab_default.interface = &default_grab_interface;
+  self->grab_default.cursor = self;
+  self->grab = &self->grab_default;
   self->screen = NULL;
   self->visible = true;
   zn_cursor_update_size(self);
